kevinarpe-papaya-testing-slf4j/src/main/java/com/googlecode/kevinarpe/papaya/logging/slf4j/SLF4JMarkerNone.java
package com.googlecode.kevinarpe.papaya.logging.slf4j;

/*
 * #%L
 * This file is part of Papaya.
 * %%
 * Copyright (C) 2013 - 2014 Kevin Connor ARPE (kevinarpe@gmail.com)
 * %%
 * Papaya is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 * 
 * GPL Classpath Exception:
 * This project is subject to the "Classpath" exception as provided in
 * the LICENSE file that accompanied this code.
 * 
 * Papaya is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 * 
 * You should have received a copy of the GNU General Public License
 * along with Papaya.  If not, see <http://www.gnu.org/licenses/>.
 * #L%
 */

import com.google.common.collect.Iterators;
import com.googlecode.kevinarpe.papaya.annotation.FullyTested;
import com.googlecode.kevinarpe.papaya.object.StatelessObject;
import org.slf4j.Marker;

import java.util.Iterator;

/**
 * @author Kevin Connor ARPE (kevinarpe@gmail.com)
 */
@FullyTested
public final class SLF4JMarkerNone
extends StatelessObject
implements Marker {

    public static final SLF4JMarkerNone INSTANCE = new SLF4JMarkerNone();

    public static final String NAME = "";

    private SLF4JMarkerNone() {
        // Empty
    }

    @Override
    public String getName() {
        return NAME;
    }

    @Override
    public void add(Marker reference) {
        // Do nothing
    }

    @Override
    public boolean remove(Marker reference) {
        return false;  // reference could not be found and removed
    }

    @Deprecated
    @Override
    public boolean hasChildren() {
        return false;
    }

    @Override
    public boolean hasReferences() {
        return false;
    }

    @Override
    public Iterator iterator() {
        Iterator x = Iterators.emptyIterator();
        return x;
    }

    @Override
    public boolean contains(Marker other) {
        return false;
    }

    @Override
    public boolean contains(String name) {
        return false;
    }
}
